refactor(libecalc): move id/name storage into EnergyUnit base class
