+++ src/libecalc/energy/energy_units/consumers.py
@@ -1,5 +1,3 @@
-from typing import Final
-
 from libecalc.energy.consumer import Consumer
 from libecalc.energy.energy_types import DieselRate, ElectricalPower, FuelGasRate, MechanicalPower
 from libecalc.energy.energy_unit import EnergyUnitId
@@ -13,19 +11,12 @@ class BaseLoad(Consumer):
     """
 
     def __init__(self, name: str, load: float, energy_unit_id: EnergyUnitId | None = None) -> None:
-        self._name = name
+        super().__init__(name, energy_unit_id)
         self._load = load
-        self._id: Final[EnergyUnitId] = energy_unit_id or BaseLoad._create_id()
 
     @classmethod
     def get_input_energy_type(cls) -> type[ElectricalPower]:
         return ElectricalPower
-
-    def get_id(self) -> EnergyUnitId:
-        return self._id
-
-    def get_name(self) -> str:
-        return self._name
 
     def get_load(self) -> float:
         return self._load
@@ -38,19 +29,12 @@ class Compressor(Consumer):
     """Compressor requiring mechanical power (driven by turbine or motor)."""
 
     def __init__(self, name: str, power: float, energy_unit_id: EnergyUnitId | None = None) -> None:
-        self._name = name
+        super().__init__(name, energy_unit_id)
         self._power = power
-        self._id: Final[EnergyUnitId] = energy_unit_id or Compressor._create_id()
 
     @classmethod
     def get_input_energy_type(cls) -> type[MechanicalPower]:
         return MechanicalPower
-
-    def get_id(self) -> EnergyUnitId:
-        return self._id
-
-    def get_name(self) -> str:
-        return self._name
 
     def get_power(self) -> float:
         return self._power
@@ -63,19 +47,12 @@ class Pump(Consumer):
     """Pump requiring mechanical power (e.g. water injection pump)."""
 
     def __init__(self, name: str, power: float, energy_unit_id: EnergyUnitId | None = None) -> None:
-        self._name = name
+        super().__init__(name, energy_unit_id)
         self._power = power
-        self._id: Final[EnergyUnitId] = energy_unit_id or Pump._create_id()
 
     @classmethod
     def get_input_energy_type(cls) -> type[MechanicalPower]:
         return MechanicalPower
-
-    def get_id(self) -> EnergyUnitId:
-        return self._id
-
-    def get_name(self) -> str:
-        return self._name
 
     def get_power(self) -> float:
         return self._power
@@ -88,19 +65,12 @@ class SampledFuelConsumer(Consumer):
     """Consumer with fuel rate from lookup (e.g. compressor+turbine as black box)."""
 
     def __init__(self, name: str, fuel_rate: float, energy_unit_id: EnergyUnitId | None = None) -> None:
-        self._name = name
+        super().__init__(name, energy_unit_id)
         self._fuel_rate = fuel_rate
-        self._id: Final[EnergyUnitId] = energy_unit_id or SampledFuelConsumer._create_id()
 
     @classmethod
     def get_input_energy_type(cls) -> type[FuelGasRate]:
         return FuelGasRate
-
-    def get_id(self) -> EnergyUnitId:
-        return self._id
-
-    def get_name(self) -> str:
-        return self._name
 
     def get_fuel_rate(self) -> float:
         return self._fuel_rate
@@ -113,19 +83,12 @@ class SampledPowerConsumer(Consumer):
     """Consumer with power demand from lookup (e.g. sampled compressor on electrical drive)."""
 
     def __init__(self, name: str, power: float, energy_unit_id: EnergyUnitId | None = None) -> None:
-        self._name = name
+        super().__init__(name, energy_unit_id)
         self._power = power
-        self._id: Final[EnergyUnitId] = energy_unit_id or SampledPowerConsumer._create_id()
 
     @classmethod
     def get_input_energy_type(cls) -> type[ElectricalPower]:
         return ElectricalPower
-
-    def get_id(self) -> EnergyUnitId:
-        return self._id
-
-    def get_name(self) -> str:
-        return self._name
 
     def get_power(self) -> float:
         return self._power
@@ -138,19 +101,12 @@ class Flare(Consumer):
     """Flare consuming fuel gas."""
 
     def __init__(self, name: str, fuel_rate: float, energy_unit_id: EnergyUnitId | None = None) -> None:
-        self._name = name
+        super().__init__(name, energy_unit_id)
         self._fuel_rate = fuel_rate
-        self._id: Final[EnergyUnitId] = energy_unit_id or Flare._create_id()
 
     @classmethod
     def get_input_energy_type(cls) -> type[FuelGasRate]:
         return FuelGasRate
-
-    def get_id(self) -> EnergyUnitId:
-        return self._id
-
-    def get_name(self) -> str:
-        return self._name
 
     def get_fuel_rate(self) -> float:
         return self._fuel_rate
@@ -163,19 +119,12 @@ class DieselConsumer(Consumer):
     """Direct diesel consumer (e.g. mobile rig, emergency generator)."""
 
     def __init__(self, name: str, rate: float, energy_unit_id: EnergyUnitId | None = None) -> None:
-        self._name = name
+        super().__init__(name, energy_unit_id)
         self._rate = rate
-        self._id: Final[EnergyUnitId] = energy_unit_id or DieselConsumer._create_id()
 
     @classmethod
     def get_input_energy_type(cls) -> type[DieselRate]:
         return DieselRate
-
-    def get_id(self) -> EnergyUnitId:
-        return self._id
-
-    def get_name(self) -> str:
-        return self._name
 
     def get_rate(self) -> float:
         return self._rate
